+++ Tests/Common/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Unity.RuntimeSceneSerialization.Tests.Playmode")]
+
+#if UNITY_EDITOR
+[assembly: InternalsVisibleTo("Unity.RuntimeSceneSerialization.Tests.Editor")]
+#endif